feat: ItemHasId availability check
