+++ client/src/resources/CurrentItemProvider.js
@@ -1,0 +1,10 @@
+export default class CurrentItemProvider {
+
+  setCurrentItem(item) {
+    this.item = item;
+  }
+
+  getCurrentItem(item) {
+    return this.item;
+  }
+}
