+++ client/src/resources/availability-checks/ItemHasIdAvailabilityCheck.js
@@ -2,22 +2,15 @@ import { inject } from 'aurelia-framework';
 import CurrentItemProvider from 'resources/CurrentItemProvider.js';
 
 @inject(CurrentItemProvider)
-export default class IdGenerator {
+export default class ItemHasIdAvailabilityCheck {
 
   constructor(currentItemProvider) {
     this.currentItemProvider = currentItemProvider;
   }
 
-  getId() {
-    return Date.now() + '-' + (Math.random() * (10 ** 9)).toFixed(0);
-  }
-
-  getIdWithCurrentItemId() {
+  async isAvailable(availabilityCheck) {
     const item = this.currentItemProvider.getCurrentItem();
-    if (!item.id) {
-      return null;
-    }
-
-    return item.id + '-' + this.getId();
+    return item && item.conf && item.conf._id;
   }
+
 }
